+++ ocaml-lsp-server/src/switch_impl_intf.mli
@@ -4,6 +4,8 @@ val capability : string * Json.t
 
 val meth : string
 
+val get_intf_impl_counterparts : string -> string list
+
 val on_request :
      params:Json.t option
   -> State.t