+++ ocaml-lsp-server/test/e2e/src/LanguageServer.ts
@@ -1,6 +1,5 @@
 import * as cp from "child_process";
 import * as os from "os";
-import * as fs from "fs";
 import * as path from "path";
 import * as rpc from "vscode-jsonrpc";